Методы вынесены в классы , вызов происходит в main
